pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "myrepetl"
version = "1.0.0"
description = "MySQL Replication ETL Tool - инструмент для репликации данных из MySQL с поддержкой трансформаций"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Tumurzakov", email = "tumurzakov@example.com"}
]
maintainers = [
    {name = "Tumurzakov", email = "tumurzakov@example.com"}
]
keywords = ["mysql", "replication", "etl", "binlog", "data-pipeline", "database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "pymysql>=1.1.2",
    "cryptography>=41.0.7",
    "mysql-replication>=1.0.9",
    "pandas>=2.1.4",
    "numpy>=1.24.3",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[project.urls]
Homepage = "https://github.com/tumurzakov/myrepetl"
Repository = "https://github.com/tumurzakov/myrepetl"
Documentation = "https://github.com/tumurzakov/myrepetl#readme"
"Bug Tracker" = "https://github.com/tumurzakov/myrepetl/issues"

[project.scripts]
myrepetl = "myrepetl.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["myrepetl*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.sql", "*.yaml", "*.yml"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
